# docs/api/strategies.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OptionsForOptions2 API",
    "version": "0.1.0",
    "description": "W-061 slice: strategy drafts. Generated by scripts/generate_openapi.py; only the /strategies paths kept."
  },
  "paths": {
    "/strategies": {
      "get": {
        "summary": "List Strategies",
        "operationId": "list_strategies_strategies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Save Draft",
        "operationId": "save_draft_strategies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/strategies/{strategy_id}": {
      "get": {
        "summary": "Get Strategy",
        "operationId": "get_strategy_strategies__strategy_id__get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Strategy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Strategy",
        "operationId": "update_strategy_strategies__strategy_id__put",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Strategy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/strategies/{strategy_id}/history": {
      "get": {
        "summary": "Strategy History",
        "operationId": "strategy_history_strategies__strategy_id__history_get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Strategy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/strategies/{strategy_id}/restore/{seq}": {
      "post": {
        "summary": "Restore Strategy",
        "operationId": "restore_strategy_strategies__strategy_id__restore__seq__post",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Strategy Id"
            }
          },
          {
            "name": "seq",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Seq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LegIn": {
        "properties": {
          "contract_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Contract Id",
            "description": "The catalogue's internal contract id (never a broker token)"
          },
          "action": {
            "type": "string",
            "enum": [
              "BUY",
              "SELL"
            ],
            "title": "Action"
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "description": "Units (lots x lot size)"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "contract_id",
          "action",
          "quantity"
        ],
        "title": "LegIn"
      },
      "StrategyIn": {
        "properties": {
          "underlying": {
            "type": "string",
            "enum": [
              "NIFTY",
              "SENSEX"
            ],
            "title": "Underlying"
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/LegIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Legs"
          },
          "rules_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rules Ref"
          },
          "risk_limits": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Risk Limits",
            "description": "name -> decimal as a string"
          },
          "preferences": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Preferences"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "underlying",
          "legs"
        ],
        "title": "StrategyIn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}